+++ Frontend/src/UserInterface/Sidebar/SidebarComponents/Sort.tsx
@@ -1,4 +1,4 @@
-import { useContext, useState } from 'react';
+import { useContext } from 'react';
 import styled from 'styled-components';
 import { GameContext } from '../../../App/GameContext';
